requirements.txt
asgiref==3.5.2
Django==4.0.4
django-filter==21.1
Pillow==9.1.1
psycopg2-binary==2.9.3
sqlparse==0.4.2
